WMPEditText.java writes content into existing node
Added note edit_text into edit_text_test.xml on server. WMPEditText will write it's content into this node. Additionally it will stop writing it's initial content into the node on entering the session as we shall make no changes to the XML as long as participants don't have connected. Therefore the WMPEditText in the prototype is stripped of all initial content, too.