EditText now reflects local document state
Registered the EditTextViewUpdater correctly to the AwarenessController. Remote changes on the document are now properly reflected in WMPEditText. Additionally a change was made when user input is propagated to the local document. This now happens in onKeyDown() instead of onTextChanged() to prevent remote changes from triggering another cefx-document update thus steering into an endless loop.