Added scripts and config files for EC2 Server deployment preperations and cleaning
As the Mobilis server configuration for the Amazon EC2 server is different from that of the local Mobilis development installment (mainly in the XMPP connection configuration) this commit adds the EC2 server configuration files and two scripts which are used before (wmp_server_config.sh) and after (local_server_config.sh) creating a running JAR for deployment on the Amazon EC2 server from Eclipse. These scripts take care of copying the necessary files and cleaning afterwareds.