Added *.bak file check to wmp_local_server_config.sh and temporarily change EC2 server information
WARNING: local_server_config.sh and wmp_server_config.sh were renamed to wmp_local_server_config.sh and wmp_ec2_server_config.sh because both were made accessable from everywhere in the filesystem which raised the need for a more specific naming convention.

Previously wmp_local_server_config.sh removed the XML setting files also if there were no bak files to replace them. We now check for the existence of the bak files before deleting the XMLs.

We also temporarily changed the XMLs for the EC2 server so that we can run the Mobilis server locally and it'll connect to the EC2 XMPP server. Note, that you may also run this configuration on the EC2 server, but it will be slower than the localhost configuration as there has to be made a DNS lookup first before being able to communicate with the XMPP server.