Quickfix (try-catch) to fix DOMException when editing an already existing XML attribute
This is a dirty workaround for ditching the infamous DOMException.INUSE_ATTRIBUTE_ERR when undoing an operation. The problem here is, that a.ownerElement has to be null, so that the attribute can be attached to the node. Otherwise the somewhat dump Android ElementImpl implementation thinks, that the attribute is already attached to another element.
So the solution would be either to use an own implementation of ElementImpl or set a.ownerElement to null (which not that easy, as there aren't any setters to achieve this.