Introduced CollabEditingService as standard mobilis service
This was a huge one. The CollabEditingService is now part of the MXA and
is accessible via the IXMPPService interface. Communication with the
WMP-Applications is done via IPC. To achieve this a load of new
interfaces and AIDLs and their implementations, as well as a lot of
serialization and Parcel stuff had to be written. 
The new CollabEditingService should be also usable by non-WMP
collaborative editing apps. The CollabEditingService handles all
collaborative editing stuff by making use of CEFX which is now directly
referenced by the MXA project, but can be switched to another framework
at any time by making the according changes to the CollabEditingService.
As of now there are still some references to CEFX outside of MXA. These
have to be removed by moving some more intelligence away from the apps
to MXA and the CollabEditingService. 

The next commit sports all necessary changes on the WMP-Project's side.