Added dynamic node creation (#13)
Dynamic node creation was a real bugger, as we could only use nodes (and
therefore WMPView IDs) which were already defined in the servers session
document before the start of the server. Now we may add and remove nodes
(and therefore WMPViews) as we like. Great achievement! :) Now WMP IDs
are finally put to good use.